backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyMySQL==1.1.1
alembic==1.13.2
python-multipart==0.0.9
APScheduler==3.10.4
reportlab==4.2.2
aiofiles==23.2.1

